+++ src/capability/system_agent.hh
@@ -17,7 +17,7 @@
 #ifndef __NUGU_SYSTEM_AGENT_H__
 #define __NUGU_SYSTEM_AGENT_H__
 
-#include "capability.hh"
+#include "clientkit/capability.hh"
 #include "capability/system_interface.hh"
 
 namespace NuguCapability {
